requirements.dev.txt
requests==2.25.1
PyYAML==6.0
python-dateutil==2.8.2
progressbar==2.5

pytest==7.2.0
